Big refactor for how multiple channels are handled by DSP functions (no more duplicating them, and they can handle a variable channel count)

+++ base/src/AzAudio/helpers.c
@@ -55,7 +55,7 @@ float linc(float x) {
 }
 
 float lanczos(float x, float radius) {
-	float c = cos(x * AZA_PI * 0.5f / radius);
+	float c = cosf(x * AZA_PI * 0.5f / radius);
 	return sinc(x) * c*c;
 }
